requirements.txt
amqp==5.3.1
anyio==4.12.1
asgiref==3.11.0
billiard==4.2.4
certifi==2026.1.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
Django==6.0.1
gunicorn==24.1.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
kombu==5.6.2
packaging==26.0
prompt_toolkit==3.0.52
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-telegram-bot==22.6
six==1.17.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uvicorn==0.40.0
vine==5.1.0
wcwidth==0.3.5
